further improvements
if Oneway gets an exception for PSV, it's not a oneway street anymore
for our intents and purposes
Added a function to determine whether a (parent) route relation is
suitable to copy ways from
Changed waysLeadingToNextStop so it now won't copy anymore from routes
where the sequence of ways is broken. I should probably add a way to
flag these routes as in need of attention.
Same with findRouteRelationWithTheLongestSequenceOfStopsInCommon()